refactor generate_input_parameter_values for unit testing
this is an example of a refactor of a component of the Taskcat class in
order to illustrate some techniques to break larger components out in
order to make them testable.

this is not a comprenhensive refactor and it's not intended to be
merged, simply as way to illustrate a process and provide some pointers.
There is a lot more work to be done here and I'd love to continue, but
will be out for a week, if someone else would like to pick this as a
starting point I would be ok with this and would love to check back and
collaborate on this after I get back.