Tidy up Docker installers
There was no real need for these to be a Python script that then awkwardly shells out instead of it just being a shell script.

Also collapses the master and develop installers into a single script.
Run the develop installer with `--develop`.

Includes a bunch of fixes for the non printf warnings that [Shellcheck](https://www.shellcheck.net/) complained about.
Left the echos as is instead of using printf because I disagree in this particular case.