Use wait logic to create taskcat S3 auto_bucket
The current implementation of S3 bucket creation plus tagging has a reproducible issue:

botocore.errorfactory.NoSuchBucket: An error occurred (NoSuchBucket)
when calling the PutBucketTagging operation: The specified bucket does
not exist

This happens because the bucket we try to tag doesn't exist yet.

The changed code uses a bucket_exists waiter to determine if the bucket
exists which fixes the above mentioned behavior.