Update README.md
`s3_regional_buckets` key is not documented. Its significance is important, because of the recommended practice when it comes to using nested stacks inside the organization repos. If the project is not uploaded to different buckets in each region, the multi-region tests fail, as the QSS3Bucket is not found in QSS3Region